refactor: 定数の切り分け
